chore: move vite config to wxt config; fix a few imports
